Seperate Python Spatial Analysis for clarity
