+++ GenerateCPCMap.py
@@ -18,11 +18,6 @@ import datetime as dt
 import numpy as np
 import matplotlib.pyplot as plt
 import matplotlib as mpl
-import shapely as shp
-from shapely.geometry import shape
-from shapely.geometry import Point
-import json
-from rtree import index
 from stravalib.client import Client
 
 def ReadCPCFile(CPCtext):
@@ -230,30 +225,3 @@ def CreateColourBar(binLims, colsHex, colorProfile):
     cb.set_label('particles per cubic centimetre')
     plt.savefig("static/colourbar_"+colorProfile+".png", dpi=300, transparent=True)
 
-
-def ReadGeoJSON(path):
-    with open(path) as hexFile:
-        data = json.load(hexFile)
-        hexagons = []
-        for feature in data['features']:
-            hexagons.append(shape(feature['geometry']))
-        return hexagons
-
-def SpatialJoin(points, polygons):
-    idx = index.Index()
-    count = -1
-
-    for poly in polygons:
-        count += 1
-        idx.insert(count, poly.hexagon.bounds)
-
-    for i, conc in enumerate(points.concs):
-        for j in idx.intersection((points.lons[i], points.lats[i])):
-            if Point(points.lons[i], points.lats[i]).within(polygons[j].hexagon):
-                polygons[j].concs.append(conc)
-                break
-
-    return polygons
-
-def Overlaps(polygon, latlng):
-    return polygon.contains(Point(latlng[0], latlng[1]))
